Prevent OOB access on illegal server response.
While parsing Xorg responses it is possible to trigger an out of
boundary read if the response does not contain enough bytes.

In case of IPv4, the padding normally prevents this, but IPv6
addresses can trigger an out of boundary read.

It takes a hostile xorg-server to reproduce this issue. If
os/access.c is adjusted to always use a length of 1, it is possible
to reproduce it and make it visible with an ASAN-compiled xhost.

Reading past the memory boundary could reveal sensitive information
to external DNS servers, because a lookup will be performed.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>
Reviewed-by: Matthieu Herrb <matthieu@herrb.eu>